Update flight_model.cfg
Changed the speeds and angle of steering authority to be closer to the real ones. The full range cannot be tweaked correctly though in the absence of a real Nose Wheel Steering system independent logic from the rudder steering system. So the following trade
-1 Modified the speed limit for gear extension : 260 kts
-2 Modified the max speed for full deflection as per NWS settings (20 kts)
Note: The NWS is supposed to be still operative until 80 kts but since this is not independent, the decrease in steering authority is too slow, so the threshold is set to 40 kts, as this is the rudder max deflection speed, in order to avoid too tight turns at high speed during taxi.
-3 The min deflection rudder at 40kts has been set to 6 degrees which is the max rudder authority angle if using the rudder only. This angle should decrease with speed as well (from 6degrees max  down to 0 degrees at 130 kts) but since NW and rudder steering cannot be separated for now this decrease from 6 to cannot happen.