[CDU] Add formatting helper (#1830)
This commit introduces a formatting helper for the CDU. With it,
developers may use special tokens in MCDU page titles, line items and
line labels in order to format different parts of a cell differently.

For example, by wrapping part of a cell in {green} and {end}, that part
of the cell will be rendered in green, and the tokens will be removed.
Similar tokens exist for all MCDU colors. Furthermore, the {small}
token can be used to force small text. Finally, a {sp} token can be
used to introduce a non-breaking space.

This work was originally done by Tyler as part of the MCDU font rework
and was split into this commit in order to isolate the change.

Co-authored-by: Tyler Knox <tyler58546@gmail.com>

Co-authored-by: Tyler Knox <tyler58546@gmail.com>