Add fs-base-ui typedefs (#1785)
SimPlane has a lot of functions that may return null,
as SimVar.GetSimVarValue is used by most of the Simplane functions.
Only functions that cannot return null (e.g. because they use
Math.max which never returns null) have an output type without null.

The Revealing Module Pattern is used a lot in Asobo's code, therefore
many of the definitions contain namespaces, which are best suited to
represent this pattern.