fix: upper ECAM showing engineering mode overlay (#2989)
* fix: upper ECAM showing engineering mode overlay

When developing the a32nx in combination with WebUI-DevKit for
debugging purposes, the upper ECAM often showed the engineering mode
overlay instead of the normal view. Whenever this happened, there were
many errors in the console stating an update function couldn't be found
on the eicas-common-display.

I traced this issue to the loading of scripts using the import-script
attribute. Scripts by default are loaded asynchronously, meaning the
declaration order isn't the loading order. In practice, usually this
doesn't lead to problems because the file system and browser cache
provides files at a quick enough rate for the order to be respected in
most cases. However, as WebUI-DevKit has cache-busting code which
removes caching altogether, the issue was more likely visible with it.
The issue did also affect regular users of the a32nx, albeit odds of
running into the issue were much smaller.

The various ECAM templates and script depend on the EICASTemplateElement
class, which is declared in A32NX_BaseAirliners.js. I've now configured
the loading such that it will always load that file first, before
loading the ECAM templates.

Loading the NXDataStore.js twice also caused errors, which I resolved
by removing one of the imports.

* Update CHANGELOG.md

Co-authored-by: Benjamin Dupont <benjozorkfr@gmail.com>
Co-authored-by: pepperoni505 <49950581+pepperoni505@users.noreply.github.com>