feat: auxiliary power unit (#3004)
* docs: apu CHANGELOG

* feat: ECAM gauge red arc can be modified

* feat: add WASM panel

* feat: first steps in integrating wasm APU

* feat: amber outside of range APU parameters

* fix: sim APU starting too slow

* chore: ignore WASM files

* feat: APU GEN volts, amperes and hertz from WASM

* feat: APU EGT rounded to nearest 5 degrees

* refactor: remove unused variables

* refactor: further simplify Lower ECAM APU update

* fix: APU available arrow and load % flickering

See the comment in A32NX_APU.js for further details.

* feat: use A32NX_APU_N instead of APU PCT RPM

Also add A32NX_APU_START_CONTACTOR_ENERGIZED
to support the related sound.

* docs: add missing A32NX simvar

* fix: start contactor disc sound not playing

* feat: show APU upper ECAM AVAIL messages

* feat: use apu avail for 10s APU page delay

* use APU avail where possible

* refactor: move APU GEN normal range logic to Rust

* feat: APU BLEED air

* fix: EXT PWR can be turned on when not available

* feat: show APU bleed pressure when ADIRS1 on NAV

* feat: APU fuel fault auto shutdown

* feat: APU inoperable status message

* fix: APU SWITCH temporarily became false

Previously I wrote that something outside of our code is setting the APU
SWITCH to 0. It turns out this is related to the FUEL_SYSTEM_VALVE
leading to the APU. When it is closed and the APU is running, MSFS will
turn the sim's APU off. By also opening the fuel valve at the same time
as starting the sim's APU we fix problems related to this "flickering".

* feat: tooltip for APU available

* fix: SHUTDOWN should be SHUT DOWN

* feat: auto shut down ECAM action

* fix: APU bleed PSI measured after the valve

* feat: APU MASTER checklist item

* fix: wrong APU sounds

* feat: APU fire emergency shut down

* feat: APU BLEED FAULT light

* fix: bleed pressure should be 0 with master off

* feat: add systems WASM build

* chore: move sound back to original position

* Integrate Rust code in a32nx

* Update A32NX/html_ui/Pages/A32NX_Core/A32NX_APU.js

Co-authored-by: Benedict Etzel <developer@beheh.de>

* docs: update CHANGELOG with APU changes

* remove unused lib.rs

* chore: cargo clippy fixes

* feat: run rust checks during CI

* fix: START sw remains lit when MASTER SW off

Before the start sequence, if you turn on MASTER SW and START SW
at the same time and then turn OFF MASTER SW, the START SW remains
lit, despite the start sequence not having begun.

Co-authored-by: Benedict Etzel <developer@beheh.de>