revert: APU FLAP OPEN indication until closed (#3768)
Based on real-world aircraft testing, Komp concluded that the FLAP OPEN
indication does in fact immediately disappear on starting to close,
instead of only disappearing when the flap is fully closed.

Note that while the flap is not fully closed, the APU data (N, EGT) will
continue to be displayed. We do this to simulate the ECB's behaviour,
which waits until the flap is fully closed before turning itself off.

Refs: b395023f7f97e8289d6e44735a09f244e3ab975f