feat(elec): replace default electrical system (#3980)
* docs: CHANGELOG

* docs: heading indentation

* feat: define infinibat and buses

* Assigned individual circuits to their electrical busses

Komps first commit

* feat: connect msfs default buses

* feat: define new buses powered by WASM module

* feat: RMPs powered by correct buses

* feat: remove APU master dependency on BUS VOLTAGE

* feat: remove APU start dependency on BUS VOLTAGE

* chore: remove unused ACPowerStateChange code

The toggling of circuit 79 is a remnant from PR #796. In PR #2119 Harry
changed circuit 79 to refer to a different element. On requesting
information Harry indicated not being aware of this code, nor of how to
test if this code is needed. Therefore I'll simply remove it and we'll
see what happens...

* feat: PFDs powered by their respective buses

* feat: NDs powered by their respective buses

* feat: ECAMs powered by their respective buses

* feat: MCDU powered by one of two buses

Currently the MCDU isn't separated per side. Once it is, the condition
for what powers the screen can be modified.

* feat: fake FADEC powered by DC ESS

This will have to be revisited when implementing the FADEC. One
shouldn't consider this reference to DC ESS valuable: it might be
powered by multiple buses or related to other things altogether.

* feat: analog BRK gauge signal source

The signals to the BRK gauge come from DC ESS powered sources. Therefore
we check for DC ESS power to be available in order to determine what
to display.

As the BRK gauge is an analog gauge, I removed the condition of it being
only displayed when there is power. There is still an issue with this:
the panel is a source of light in a C&D cockpit while it shouldn't be.

* feat: ISIS powered by the appropriate buses

* chore: get rid of !!s

* feat: change NavSystem to DC ESS

It's unclear what the NavSystem actually does. It might be partially
dead code for our project. The isElectricityAvailable function doesn't
exist within the repository.

Picking the DC ESS bus, because other than the HOT BUSes it's the most
frequently available DC bus.

* feat: XPDR 1 powered by AC ESS SHED

* feat: clock powered by DC ESS or DC HOT 1

* feat: rudder trim powered by DC 2

* feat: MCDUs powered by AC ESS SHED and AC 2

* feat: DCDUs powered by DC 1 and 2

* refactor: infinite battery

* chore: remove unused code

* feat: lights powered by bus

* feat: lower overhead panel lights power

* feat: RAT & EMER GEN light DC ESS

* feat: always power the panel

* feat: BAT OFF light powered by DC BAT bus

* feat: one second BCL startup delay

* feat: maintenance panel buses

* feat: fire button buses

* feat: various center panel lights

* feat: auto land and side stick priority lights

* fix: circuit connections to new buses

* feat: power TERR ON ND with AC BUS 1

* feat: landing gear lights buses

* feat: power BRK FAN lights with AC BUS 2

* feat: power AUTO BRK lights with AC BUS 1

* feat: power cockpit door light with AC BUS 2

* feat: DCDU power available

* feat: power EWD with AC ESS bus

* feat: power SD with AC 2 bus

* chore: remove unused L:APU_GEN_ONLINE

* chore: remove unused EXTERNAL POWER variables

* feat: main backlighting uses AC BUS 1

* feat: brake gauge (again) powered by DC ESS

* fix: some lights are flashing

* feat: AC 1 and AC ESS SHED power FCU buttons

* refactor: powering of DisplayUnit

* feat: power MCDU buttons with AC BUS 1

* feat: power ECAM button lights

* feat: power DCDU buttons with AC BUS 1

* feat: tiller and throttle emissive

* feat: power ISIS button lights with AC BUS 1

* feat: power brake gauge lights with AC BUS 1

* feat: reading lights connect to DC BUS 1

* docs: lights

* feat: pedestal flood light powered by DC 1

* docs: light names

* feat: glareshield lights

* feat: console/floor lights powered by DC 1 and 2

* feat: dome lights powered by DC 2 and DC ESS

Waiting for emissive to be separated into captain and F/O side before
adapting that.

* feat: power RMP

* feat: power audio control panels

* feat: power XPDR

* fix: alt selector not assigned to bus

* chore: remove unused A320_NEO_BRK

* docs: refer from systems.cfg to panel.xml

* feat: power FCU with DC ESS or DC 2

* feat: dome lights powered by DC 2 and DC ESS

* chore: tabs to spaces

* chore: remove unused CSS

* docs: changelog

* refactor: move A320 specific template concerns

* chore: remove "authors"

Git already stores author information for us.

* feat: checklists

* refactor: remove duplicate light test condition

* fix: infinibat on flight start not always ON

* feat: COM2 powered by DC2, COM3 by DC1

Co-authored-by: komp1821 <71044068+komp1821@users.noreply.github.com>