feat(fpm, lnav): refactoring of flight path building / rendering. add support for VM legs (#5362)
* feat(fpm, lnav): refactoring of flight path building / rendering. add support for VM legs

* fix: apply suggestions from code review

* fix(ND): missing transitions, def Tf->TF

* fix: VM leg heading not taken into account in flight plan syncing and multipleLegGeometry

* fix: do not sequence discontinuity if there is a VM leg after

* fix: getActiveLegPathGeometry VM leg generation

* fix: respect when vector is last wp in segment

* fix: mag var issue on VM legs

* fix(nd): recompute geometry every 2 seconds

* fix(fmgc): incorrect spacing, discont. msg when clr attempt, modification of tmp. fpln fixed

* fix: VM guidance mag var

* fix: use magnetic heading for to wp info

* fix: draw transitions >180 degrees

* fix: round heading for bug

For some reason it sometimes becomes a non-integral value...

* fix(nd): adjust map masks

Should be very close to realistic. Will need a final pass of all
geometry later for consistency.

* feat(nd): constraints display

* fix(nd): wind flicker 0 -> 360 -> 0

* fix(nd): alt constraint circle

Alt constrained waypoints are displayed with a circle around them. The
circle should be white when the constraint is not taken into account/NAV
is not engaged, magenta when it will be met, and amber when it will be
missed. The met/missed predictions are not available yet (VNAV), so for
now we just display them in white.

* fix(cdu): cumulative distance/time calculations

Co-authored-by: 2hwk <reasonstouninstall@tutanota.com>
Co-authored-by: Michael Corcoran <tracer@outlook.co.nz>