fix: NXDataStore cannot store numbers (#5741)
* fix: NXDataStore cannot store numbers

The underlying Coherent datastore errors out when anything other than a
string is set.

* fix: refactor NXDataStore

usePersistentPropertyWithDefault was not needed (usePersistent property could already do what it did).
There was a bug in usePersistentProperty which stopped it from ever setting the default in the datastore... and it is bound to be racy if you rely on it across instruments etc. anyway... removed.

provide userPersistentNumberProperty for convenience.

Added overloads to avoid possibly undefined issues for callers when they provide a default. Required eslintrc rules to be changed as they were not typescript aware.

* fix(efb): make default baro setting auto

in.Hg was inconsistent to metric units being the default everywhere else

* fix: bug with empty keys

* fix: auto-brightness moves brightness slider

* fix: restore brightness setting when auto turned off

* fix: stfu linter

* fix: racy brightness

* style: fix lint

Co-authored-by: 2hwk <15316958+2hwk@users.noreply.github.com>
Co-authored-by: Benjamin Dupont <benjozorkfr@gmail.com>