feat(fwc): add takeoff surveillance 1 warnings and various other improvements (#7652)
* chore(msfssdk): add missing DebounceTimer

* fix: clean up adr altitude labels and remove isa deviation

I have not implemented the baro corrected alt labels correctly yet as they still rely on the sim altimeters => they change to pressure alt when STD is selected, but at least the proper labels are there so other subsystems canhook up to the proper labels and end up correct in future.

There are things that probably ought to be using pressure alt rather than baro corrected alt, but I didn't go into researching all of that. That's a result of having the ADR labels messed up since the beginning.

* fix(cdu): scratchpad bug

* fix(fwc): add method to read mtrig

* feat(fwc): implement pulse node

* fix(fwc): split aural and text warnings

Not perfect, but at least better than before and enables proper implementation of the warnings in this branch..

* fix(fwc): fix to config/clr/recall

Bit closer to proper logic so that the new warnings can function properly...

* feat(fwc): takeoff surveillance and associated warnings

* fix(rs): actually read pressure alt... oop

* fix(sd): don't init with adr/ir source 0

This actually results in creating simvars with index 0 which we don't really want.

* fix(fmgc): correct v2 min

* fix: code review

Co-authored-by: Eearslya Sleiarion <eearslya@gmail.com>
Co-authored-by: BBK <22713769+BlueberryKing@users.noreply.github.com>

* fix(fm): refactor arinc vars now that we can write them

* doc: fm simvars

---------

Co-authored-by: Eearslya Sleiarion <eearslya@gmail.com>
Co-authored-by: BBK <22713769+BlueberryKing@users.noreply.github.com>