fix(efb): simbridge remote setting fix (#7989)
* Fixing Remote SimBridge IP not resetting after selecting 'My PC'

Added extra logic to the getSimBridgeUrl() function to default to 'localhost' for the IP address when the SimBridge is in local mode rather than remote. This way, a user switching between 'This PC' and 'Remote PC' under the FlyPad Sim Options doesn't need to manually change the IP to 'localhost' for local SimBridge to work and can retain the saved IP for switching back.

* Fixing Layout

Minor Change

* fix(efb): added quick reconnection after setting change

Added subscribe to the CONFIG_SIMBRIDGE_REMOTE in order to check for server availability once this setting is changed rather than relying on the interval

* Updating CHANGELOG.md

Updating CHANGELOG.md

---------

Co-authored-by: Saschl <19493808+Saschl@users.noreply.github.com>
Co-authored-by: Benjamin Dupont <benjozorkfr@gmail.com>